// src/app/users/model/UserRole.ts
export enum UserRole {
  SHOW_READ = 'SHOW_READ',
  SHOW_WRITE = 'SHOW_WRITE',
  SHOW_ANALYTICS = 'SHOW_ANALYTICS',
  EPISODES_READ = 'EPISODES_READ',
  EPISODES_WRITE = 'EPISODES_WRITE',
  NEWS_READ = 'NEWS_READ',
  NEWS_WRITE = 'NEWS_WRITE',
  SUBSCRIBERS_READ = 'SUBSCRIBERS_READ',
  SUBSCRIBERS_WRITE = 'SUBSCRIBERS_WRITE',
  SUPPORT_READ = 'SUPPORT_READ',
  SUPPORT_WRITE = 'SUPPORT_WRITE',
  CONTACTS_READ = 'CONTACTS_READ',
  CONTACTS_WRITE = 'CONTACTS_WRITE',
  USER_READ = 'USER_READ',
  USER_WRITE = 'USER_WRITE',
  PROFILE_READ = 'PROFILE_READ',
  PROFILE_WRITE = 'PROFILE_WRITE',
  ROLES_UPDATE = 'ROLES_UPDATE',
  NOTIFICATIONS_READ = 'NOTIFICATIONS_READ',
  SETTINGS_READ = 'SETTINGS_READ',
  SETTINGS_WRITE = 'SETTINGS_WRITE'
}